feat: Add boolean and comparison tokens
- Added support for the `BOOL` token.
- Added support for the `EQUAL` and `NEQUAL` tokens.

This commit introduces two new token types: `BOOL` and `EQUAL` and `NEQUAL`. The `BOOL` token is used to represent boolean values, which are either true or false. The `EQUAL` and `NEQUAL` tokens are used to represent the comparison operators "==" and "!=" respectively.

These changes extend the language's capabilities by allowing for the creation of more complex expressions, which can be used to perform more sophisticated calculations and comparisons.